refactor: migrate media to content layer
