+++ content/types/index.ts
@@ -1,0 +1,15 @@
+export * from './page'
+export * from './media'
+
+/**
+ * Utility type to allow known string literals plus any other string
+ *
+ * @usage
+ *
+ * ```ts
+ * type MyType = LooseUnion<'option1' | 'option2'>
+ * const a: MyType = 'option1' // valid
+ * const b: MyType = 'custom'  // also valid
+ * ```
+ */
+export type LooseUnion<T extends string> = T | (string & {})
